Broke down checkpoints/lifecycle.rs to submodules
